interconnect: allow non-sleeping writes
Having write-through behaviour for caches seems a necessary evil,
without an "always consistent" dram state, we cannot even boot.

For this reason it makes sense to add non-sleeping writes, which
will happen with no delay. This feature will allow us to implement
a cache model with a write-back policy, making non-sleeping writes
to memory just to keep the system working.

Signed-off-by: Emilio G. Cota <cota@braap.org>